Moved validation of remote data folder to common
